backend/api_gateway/requirements.txt
Flask==3.0.3
requests==2.29.0
redis==5.1.1
pyjwt==2.9.0
flask-cors==5.0.0
